# Backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API For Client",
    "description": ""
  },
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "authorization"
    }
  },
  "host": "localhost:4000",
  "basePath": "/api",
  "schemes": ["http", "https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "servers": [
    "http://localhost:4000",
    "https://gift-circle-server.onrender.com"
  ],
...
  "paths": {
...
    "/gift/{id}": {
      "get": {
        "description": "",
        "tags": ["Gift"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/gift/Get/EventGifts/{id}": {
      "get": {
        "description": "",
        "tags": ["Gift"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/gift/Get/EventGiftsTrans/{id}": {
      "get": {
        "description": "",
        "tags": ["Gift"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/gift/Get/PurchasedBy/{id}": {
      "get": {
        "description": "",
        "tags": ["Gift"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
    "/gift/create": {
      "post": {
        "description": "",
        "tags": ["Gift"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "type": "array",
            "schema": { "$ref": "#/definitions/schema/Gift" }
          }
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/gift/createMany": {
      "post": {
        "description": "",
        "tags": ["Gift"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "type": "array",
            "schema": { "$ref": "#/definitions/schema/Gift" }
          }
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/gift/Buy": {
      "post": {
        "description": "",
        "tags": ["Gift"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "type": "array",
            "schema": { "$ref": "#/definitions/schema/BuyGift" }
          }
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/gift/EnableContribution": {
      "put": {
        "description": "",
        "tags": ["Gift"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "type": "array",
            "schema": { "$ref": "#/definitions/schema/EnableContribution" }
          }
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/gift/EnableContribution/{id}": {
      "put": {
        "description": "",
        "tags": ["Gift"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/gift//{id}": {
      "delete": {
        "description": "",
        "tags": ["Gift"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
    "/fundRaising/{id}": {
      "get": {
        "description": "",
        "tags": ["FundRaising"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/fundRaising/GetFundDonors/{id}": {
      "get": {
        "description": "",
        "tags": ["FundRaising"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/fundRaising/create": {
      "post": {
        "description": "",
        "tags": ["FundRaising"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "type": "array",
            "schema": { "$ref": "#/definitions/schema/FundRaising" }
          }
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/fundRaising/Donate": {
      "post": {
        "description": "",
        "tags": ["FundRaising"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "type": "array",
            "schema": { "$ref": "#/definitions/schema/FundDonate" }
          }
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/fundRaising/UpdateStatus": {
      "put": {
        "description": "",
        "tags": ["FundRaising"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "type": "object",
            "schema": {
              "properties": {
                "status": {
                  "type": "string"
                },
                "id": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/fundRaising/UpdateAmount": {
      "put": {
        "description": "",
        "tags": ["FundRaising"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "type": "object",
            "schema": {
              "properties": {
                "amount": {
                  "type": "number"
                },
                "id": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
  },
...
}